✨ Add mediator for insert new operation

+++ Modules/Transfers/Modules.Transfers.Core/Dtos/CreateTransferDto.cs
@@ -1,3 +1,3 @@
 namespace Modules.Transfers.Core.Dtos;
 
-public record CreateTransferDto(string Iban, string Beneficiary, float Amount, string Description, Guid CardId);
+public record CreateTransferDto(string Iban, string Beneficiary, decimal Amount, string Description, Guid CardId);
