+++ Modules/Transfers/Modules.Transfers.DataAccess/Entities/TransferEntity.cs
@@ -6,7 +6,7 @@ public record TransferEntity
     public required string Iban { get; init; }
     public required string Beneficiary { get; init; }
     public DateTime Date { get; init; }
-    public float Amount { get; init; }
+    public decimal Amount { get; init; }
     public required string Description { get; init; }
     public Guid CardId { get; init; }
 }
